Update setup.py file as not implemented
This commit has two main parts:
1. Update of the setup.py file which now raises a NotImplemented error
2. Test the long message format with the "git commit" command which opens up
the default text editor prespecified for git using "git config core.editor"
as "vim". After "git commit" VIM is opened up so that I can write this
commit message and then, when I save and exit this file, the commit is
executed with the contents of this message.