Implement fast sink detection
1. The structure of the directory is now changed with a new subdirectory
./addem/helpers/
2. The module sinks.pyx/.so in the helpers directory contain a fast
implementation of sink detection achieved with Cythonized code. This was
faster than the vectorized Python code by almost 60 times!
The pozo example now gets done in just about 2 seconds of which most of
the time is spent in loading the dataset, and not in the sink detection
routine.
3. The simepl Cythonized code for sink detection passes the test case
implemented in tests/test_flows.py.