Update sink filling routine to multi-pixel sinks
Using the example from
http://spatial-analyst.net/ILWIS/htm/ilwisapp/fill_sinks_algorithm.htm
1. The sink filling routin now easilz detects single-pixel sinks and
fills them first.
2. As a second step, it detects multi-pixle sinks and detects the
catchment area for that depression, and then estimates the value with
which the identified sink(s) should be filled.

This routine is in its first draft now, nad has to be optimized. I have
not tested it for the Pozo dataset yet. Have only tried it with the test
example as given in the above URL.