Change sink filling from recursion to while-loop
1. The earlier version of the sink filling routine for multi pixels
relied on a recursive function call, which soon reached the recursion
limit for the test data case of the Pozo catchment.
2. I changed this routine to achieve the same functionality by using
while loops and appropriate stopping criteria.
3. So far, the multi pixel sink detection part works completely fine.
4. However, the multi-pixel catchment detection given a particular
multi-pixel sink fails to detect the entire catchment. This was tried
out with the 9x9 elevation model test case in ./test/test_helpers.py.

This routine needs to be corrected and then checked with the Pozo
dataset to see how it scales with large datasets.