Improved sink and catchment detection
1. Working with the Pozo test case, we found that:
    a) there might be catchments with multiple sinks in them,
    b) the fill value of a sink is dependent solely on the catchment,
    c) it is more optimal to identify all sinks within a catchment and
    then determine the single fill value for all of them in one shot.
2. As of now, when a multi-pixel sink is detected, the routine simply
calls the catchment detection function without bothering to identify the
full sink.
3. Once the catchment is identified, the 'sinks_in_catchment' function
is called which identifies all sinks within the catchment borders,
including the initial sink which had started this whole process.
4. Further work is left involving the optimization of the catchment
detection routine, and the revision of the border detection of a
catchment which is critical to identifying the fill value for all sinks
within a catchment.