Identify sinks with sweep approach
Multi-pixel sinks are identified by sweeping the DEM in two steps:
1. Identify single pixels which are:
    1.a. not at the border
    1.b. not a missing value
    1.c. do not have a way down via its neighbours
2. This set of single pixels are potentially parts of multi-pixel sinks.
Now, a second sweep of the DEm is done to:
    2.a. Identify the plateau regions associated with the single pixels
    identified above.
    2.b. Keep the plateau region as a sink only if:
        2.b.i.  the plateau is not at border
        2.b.ii. there is no way down from the plateau

These ideas are now implemented successfully in SINKS.PYX and the next
step is to go through the identified multipixel sinks and get their
respective catchments and obtain the respective fill values for each
sink.

There is also the issue of whether to update the DEM while the
catchments are being identified or to proceed in a step-wise, modular
way.