First draft of detecting sink catchments
Now that the sink detection setup does not falsely identify hill-tops
as sinks, we revisit the catchment detection algorithm and refine it
from its previous version. The goals are:

1. To be able to uniquely identify one catchment per sink -- otherwise
there is an inherent logical contradiction since once catchment cannot
have more than one sink in it, by construction.
2. Write a code that is entirelz based upon NumPy arrays (as used by
Cython) and does not rely on Pythonic objects such as tuples or lists.
3. Identify catchments and fill sinks on the fly, and then include the
entire steup in a WHILE loop which keeps iterating this entire procedure
(right from the sink detection step) until there are no more sinks to be
found in the DEM raster.