Redefine approach to sink filling
1. Catchment detection is not essential for sink filling. An iterative
approach where a sink is detected, filled with minimu of neighbours and
then re-checked to see if it is still a sink is now the new goal.
2. This is achieved in most toolboxes of using some variation of the
flood-fill appraoch from image processing.
3. The TopoToolbox uses the Matlab 'imfill' function which is
essentially a flood-fill appraoch.

Till now, the main point where I was stuck at was how to choose the fill
value for a detected multi-pixel sink. Now, the solution is simpler, as
the fill value is simply the minimum of the neighbours bordering the
detected multi-pixel. However, this then requires an iterative 'while'
loop which then checks, after sink filling, whether that region is still
a sink or not.

This is goign to be implemenedt in the next sessions.