Implement floodfill successfully
Sink filling is now implemented successfully using the flood-fill
approach. It completes the Pozo test dataset in 285 secs. This is the
best optimisation I see possible right now using Cython, where the code
is running on almost C-like run speeds.

Likely, the code can be made faster with parralelization, but that will
be done at a later time, once the entire module is up and running.