Solve problem with addem.__version__
Since setuptools does not manage to import __version__ into the
namespace ont he import of the 'addem' package, we use a trick bz import
addem in the setup.py file and then using addem.__version__ (as
mentioned in __init__.py) to specify the version for setuptools.

This way the version number does not have to be changed at more than one
place in the directory structure.